+++ documentation/RELEASE_NOTES.md
@@ -1,6 +1,7 @@
 # Release 1.X.Y (04/DD/2024)
 - Add ability to set server queue size for image viewer, which can reduce
   number of dropped frames at higher image rates
+- Add ability to display moving average images
 
 # Release 1.12.4
 - Fix payload start for LZ4 and BSLZ4 compression codecs